requirements.txt
ccxt==2.6.85
pandas==2.0.3
numpy==1.24.3
python-dotenv
requests
pytest==7.4.0
aioconsole
scikit-learn